BIG UPDATE!
over 30 warnings and bugs fixed.
commented out refrences to a "Muscle Curse" (only thing left over was the pulsing animation, I might reuse this for later endevors)
FIXED THE JIGGLE PHYSICS! this took a bit, but I am very happy it works again.
renamed some nodes and scripts to help with readability.
fixed the issue where the console floods with errors when breasts get too big.
fixed issue where the background was not aligned with the viewport